Add volume value to the Video Player volume

+++ Components/ProgressBar/Models/ProgressBarSeekButtonInfo.cs
@@ -2,9 +2,9 @@
 {
 	public class ProgressBarSeekButtonInfo
 	{
-		public int Left { get; set; }
-		public int Top { get; set; }
-		public int Width { get; set; }
-		public int Height { get; set; }
+		public double Left { get; set; }
+		public double Top { get; set; }
+		public double Width { get; set; }
+		public double Height { get; set; }
 	}
 }
